backend/requirements.txt
fastapi
uvicorn[standard]
pymysql
python-dotenv
groq
pdfplumber
apscheduler
requests
cryptography